Rework 87bb767's fix It didn't cause any issue but I didn't like the ever growing quest list. - `QuestList` is now a `ConcurrentDictionary` using value as the index to send to clients. - Added `AvailableQuestIndexes` to keep track of removed indexes that can be reused. If empty, quests are added to `QuestList` with `QuestList.Count` as the value. Reusing indexes is supported by the client and prevents the ever growing quest list. - Made `QuestListFinished` private and added a bunch of methods to access it using a lock to improve thread safety. Originally the server could crash during a daily / weekly / monthly quest tick. - Modified packet libs to correctly send null for unused indexes. Added more efficient ways to update quests. Removed duplicate methods. - Removed `NullQuest` class, `QuestList` wrapper, `QuestLock` object from last commit.

+++ GameServer/packets/Server/PacketLib171.cs
@@ -313,12 +313,11 @@ namespace DOL.GS.PacketHandler
 			}
 		}
 
-		protected override void SendQuestPacket(AbstractQuest quest, int index)
+		protected override void SendQuestPacket(AbstractQuest quest, byte index)
 		{
 			using (GSTCPPacketOut pak = new GSTCPPacketOut(GetPacketCode(eServerPackets.QuestEntry)))
 			{
-
-				pak.WriteByte((byte)index);
+				pak.WriteByte(index);
 				if (quest.Step <= 0)
 				{
 					pak.WriteByte(0);
